reorganize internals
