+++ mock/src/crypto/mod.rs
@@ -1,7 +1,9 @@
 mod keychain;
+mod signable;
 mod signature;
 mod wrappers;
 
 pub use keychain::Keychain;
+pub use signable::Signable;
 pub use signature::{PartialMultisignature, Signature};
 pub use wrappers::BadSignatureWrapper;
