+++ mock/src/lib.rs
@@ -2,14 +2,10 @@ mod crypto;
 mod dataio;
 mod hasher;
 mod network;
-mod router;
-mod signable;
 mod spawner;
 
-pub use crypto::{BadSignatureWrapper, Keychain, PartialMultisignature, Signature};
+pub use crypto::{BadSignatureWrapper, Keychain, PartialMultisignature, Signable, Signature};
 pub use dataio::{Data, DataProvider, FinalizationHandler};
 pub use hasher::{Hash64, Hasher64};
-pub use network::{Network, NetworkReceiver, NetworkSender};
-pub use router::{NetworkHook, Peer, Router};
-pub use signable::Signable;
+pub use network::{Network, NetworkHook, NetworkReceiver, NetworkSender, Peer, Router};
 pub use spawner::Spawner;
